Fix web imports (#45)
* WIP

* Consolidate if statements and shorten get calls

* Change property syntax

* Change the way the template file is set in OpagCGI

Instead of setting the template file through a function call before
printing, it is instead just passed to the print function as an
argument.

* Remove true/false variables

* Maybe fixed imports

* Fix template issues

+++ py/occampy3/wrappers/sbm_manager.py
@@ -1,6 +1,7 @@
 from manager import Manager
 from model import Model
 from occam import SBMManager as SBMManager_cpp
+from variable_list import VariableList
 
 
 class SBMManager(Manager):
@@ -18,7 +19,7 @@ class SBMManager(Manager):
         return Model(self._ref.makeSbModel(model_type, make_project))
 
     @property
-    def test_sample_size(self) -> double:
+    def test_sample_size(self) -> float:
         return self._ref.getTestSampleSz()
 
     @property
